Create header files motor and pins
