+++ HP-manouvering.code-workspace
@@ -3,5 +3,8 @@
 		{
 			"path": "."
 		}
-	]
+	],
+	"settings": {
+		"editor.tabSize": 2
+	}
 }
