+++ src/Robosub_testsetup.cpp
@@ -9,16 +9,18 @@
  *  Rutger Janssen
  *
  * Hogeschool Utrecht
- * Date: 23-09-2024
+ * Date: 24-09-2024
  *
- * Version: 2.1.0
+ * Version: 2.3.0
  * 
  * To Log the serialprints to textfile
  *    1. Open new platform io terminal
  *    2. type command 'pio device monitor > output_logs.txt'
  */
 
-#include "main.h" // Main header file
+#include "main.h" // Main header 
+#include "pins.h"
+#include "motor.h" // Motor header
 
 #define DEBUG // (un)comment to toggle (Serial) DEBUG mode 
 // #define DEBUG_VERNIER
@@ -29,21 +31,17 @@
 
 #define LCD 1 // Toggle LCD 0 to 1
 
+systemState currentState; // class storing the current system state
 enum testPrograms testProgram; // which testprogram to run
-enum direction_t direction = Forward;
 
-/* PIN DEFINTIONS */
-uint8_t VOLT_PIN = A0; // Define Voltage control
-uint8_t AMP_PIN = A1;  // Define Amperage control
+/* Button pin array */
+const uint8_t BUTTON_ARRAY[NUM_BUTTONS] = 
+  { BTN_YELLOW_PIN,
+    BTN_GREEN_PIN,
+    BTN_BLUE_PIN }; 
 
-uint8_t ESC_PIN = 3; // Define ESC control pin
-
-uint8_t VERNIER_PIN = A3; // Define Vernier analog read
-
-const uint8_t BUTTON_PINS[NUM_BUTTONS] = {6, 5, 4}; // Define ESC control pin D6 D5 D4
 Bounce *buttons = new Bounce[NUM_BUTTONS];          // Initiate 3 Bounce objects
 bool buttonStates[NUM_BUTTONS] = {false};           // bool array storing the buttonStates
-bool *pButtonStates = &buttonStates[0];             // define pointer, pointing to zeroth element of buttonStates array
 
 /* ADC Calibration values */
 float ADC_V_Step = 0.01852;
@@ -56,29 +54,6 @@ LiquidCrystal_I2C lcd(LCD_addr, LCD_COLS, LCD_ROWS); // set the LCD address to L
 
 char const *rowOneLCD = "               "; // const pointer to a char array containing linezero of LCD
 char const *rowTwoLCD = "               "; // textrow one of LCD
-
-/* Motor configuration */
-Servo esc; // Create a Servo object
-
-// Store the state of motor test program
-// program A,       B,       C
-uint8_t motorTestState[3] = {NEUTRAL, NEUTRAL, 0};
-
-// Create the 3 timers for each motortestprogram
-TimerEvent timer_motor_test_a;
-TimerEvent timer_motor_test_b;
-TimerEvent timer_motor_test_c;
-
-TimerEvent timer_motor_test_d;
-TimerEvent timer_motor_test_e;
-
-uint8_t timer_expired = 0;
-bool continuous_motor_test = true;
-
-uint32_t wait = 0;
-
-uint32_t micros_prog_a = MTR_NEUTRAL;
-uint32_t micros_prog_b = MTR_NEUTRAL;
 
 /* Measurement data storage*/
 MEASUREMENT data;           // measurement data
@@ -104,16 +79,16 @@ void setup()
   pinMode(AMP_PIN, INPUT);
   pinMode(VERNIER_PIN, INPUT); // Vernier pin as input
 
-  pinMode(BUTTON_PINS[0], INPUT_PULLUP);
-  pinMode(BUTTON_PINS[1], INPUT_PULLUP);
-  pinMode(BUTTON_PINS[2], INPUT_PULLUP);
+  pinMode(BUTTON_ARRAY[0], INPUT_PULLUP);
+  pinMode(BUTTON_ARRAY[1], INPUT_PULLUP);
+  pinMode(BUTTON_ARRAY[2], INPUT_PULLUP);
 
   pinMode(ESC_PIN, OUTPUT);
 
   // After setting up the button, setup debouncer
   for (int i = 0; i < NUM_BUTTONS; i++)
   {
-    buttons[i].attach(BUTTON_PINS[i], INPUT_PULLUP); // setup the bounce instance for the current button
+    buttons[i].attach(BUTTON_ARRAY[i], INPUT_PULLUP); // setup the bounce instance for the current button
     buttons[i].interval(25);                         // debounce interval in ms
   }
 
@@ -135,13 +110,33 @@ void setup()
 void loop()
 {
   lastReadTime = millis();
+  bool *pButtonStates = &buttonStates[0]; // define pointer, pointing to zeroth element of buttonStates array
+
   handleButtons(pButtonStates);
 
   calcPower(pData);
   motorTest(testProgram);
 
 }
+/*
+  Function: handle button presses
+  Parameters: pState, pointer to i'th index of buttonstatearray
+ */
+void handleButtons(bool *pState)
+{
+  // for the NUM_BUTTONS increase i and state pointer
+  for (int i = 0; i < NUM_BUTTONS; pState++, i++)
+  {
+    buttons[i].update(); // Update the Bounce instance
 
+    *pState = buttons[i].fell(); // change right value of this button state
+
+#ifdef DEBUG
+    if (buttons[i].fell())
+      Serial.println((String) "button: " + i + " pressed\t state: " + *pState);
+#endif
+  }
+}
 /*
   Function: Calibrate the shunt for Voltage and Current
  */
@@ -281,6 +276,7 @@ testPrograms selectProgram(void)
 {
   testPrograms selectedProg = A;
   bool progSelected = false;
+  bool *pButtonStates = &buttonStates[0]; // define pointer, pointing to zeroth element of buttonStates array
 
   #ifdef DEBUG
     Serial.println("SelectProgram()");
@@ -340,39 +336,6 @@ testPrograms selectProgram(void)
   }
 
   return selectedProg;
-}
-/*
-  Function: InitMotor
-  Puts motor to MTR_NEUTRAL µs posittion
-  Parameters: void
- */
-void initMotor(void)
-{
-  esc.attach(ESC_PIN); // Attach the ESC to the specified pin
-  delay(20);
-
-  esc.writeMicroseconds(MTR_NEUTRAL); // Send a signal to the ESC to arm it
-  delay(MTR_STARTUP_DELAY_MS);
-}
-
-/*
-  Function: handle button presses
-  Parameters: pState, pointer to i'th index of buttonstatearray
- */
-void handleButtons(bool *pState)
-{
-  // for the NUM_BUTTONS increase i and state pointer
-  for (int i = 0; i < NUM_BUTTONS; pState++, i++)
-  {
-    buttons[i].update(); // Update the Bounce instance
-
-    *pState = buttons[i].fell(); // change right value of this button state
-
-#ifdef DEBUG
-    if (buttons[i].fell())
-      Serial.println((String) "button: " + i + " pressed\t state: " + *pState);
-#endif
-  }
 }
 
 /*
@@ -518,225 +481,6 @@ void output2Serial(PMEASUREMENT p)
     Serial.print(",");
     Serial.println(p->power, 2);
   }
-}
-
-/*
-  Function: motorTest
-  Deze functie laat de motor door CYCLES standen lopen. DUR_PROG_A, DUR_PROG_B
-  Parameters: enum testPrograms prog, which tesprogram to run
-  Deps: TimerEvent.h
- */
-void motorTest(enum testPrograms prog)
-{
-  currentState = systemState::Testing; // put system to Testing
-
-  #ifdef DEBUG
-    Serial.println((String) "Testing motorprogram:" + (int)prog);
-  #endif
-
-  #if defined(LCD) && (LCD == 1)
-    lcd.clear();
-    lcd.print("Testing ;)");
-  #endif
-
-  switch (prog)
-  {
-  case A:
-    initMotor();
-    timer_motor_test_a.set(DUR_PROG_A, prog_a_timer_handler); // Set the timer
-    //       Laat de motor continue harder draaien, duurt DUR_PROG_A msecs*/
-    continuous_motor_test = true;
-    while (continuous_motor_test) // While loop gets played as long as continuous_motor_test is true
-    {
-      timer_motor_test_a.update(); // Update the timer
-
-      // Put the vernier sensor read func here (can be another timer if needed)
-      readVernier();        // force [N]
-      calcPower(pData);     // motor [A] & [V]
-      output2Serial(pData); // write data to Serial
-
-      if (timer_expired >= CYCLES) // Check if the loop has been played 500 times
-      {
-        prog = B;
-        continuous_motor_test = false;      // Set the bool to false to stop the while loop
-        timer_expired = 0;                  // Reset timer_expired
-        esc.writeMicroseconds(MTR_NEUTRAL); // Set the motor to 0 RPM
-      }
-    }
-
-    break; /* Program A */
-
-    case B:
-    initMotor();
-    timer_motor_test_b.set(DUR_PROG_A, prog_b_timer_handler);
-
-    while(1)
-    {
-      timer_motor_test_b.update(); // Update the timer
-
-      // Put the vernier sensor read func here (can be another timer if needed)
-      readVernier();        // force [N]
-      calcPower(pData);     // motor [A] & [V]
-      output2Serial(pData); // write data to Serial
-    }
-    break;
-
-  default:
-#ifdef DEBUG
-    Serial.println("ERR: Verkeerd motor test programma doorgegeven");
-#endif
-    break;
-  }
-}
-
-/*
-  Function: prog_a_timer_handler
-
-  Parameters: void
-*/
-
-void prog_a_timer_handler(void)
-{
-#ifdef DEBUG_MOTOR
-  Serial.println((String) "wait is at" + wait);
-#endif
-
-  switch (motorTestState[A])
-  {
-  case NEUTRAL: // 0
-
-    if (wait < 5)
-      esc.writeMicroseconds(MTR_NEUTRAL);
-
-    wait += 1;
-
-    if (wait >= WAIT_TIME)
-    {
-      motorTestState[A] = ADDING; // Next phases is increasing speed
-      wait = 0;
-    }
-    break;
-
-  case ADDING: // 1
-#ifdef DEBUG_MOTOR
-    Serial.println("Reached case Adding");
-#endif
-
-    if (micros_prog_a >= MTR_MAX_ANTICLOCKWISE)
-    {
-      wait += 1;
-      if (wait >= WAIT_TIME)
-      {
-        motorTestState[A] = SUBTRACTING;
-        wait = 0;
-      }
-    }
-    else
-    {
-
-      esc.writeMicroseconds(micros_prog_a = micros_prog_a + MTR_INCREMENT);
-#ifdef DEBUG_MOTOR
-      Serial.println((String) "micros_prog_a = " + micros_prog_a);
-#endif
-    }
-    break; // End of ADDING
-
-  case SUBTRACTING: // 2
-#ifdef DEBUG_MOTOR
-    Serial.println("Reached case Subtracting");
-#endif
-
-    if (micros_prog_a == MTR_MIN_CLOCKWISE)
-    {
-      wait += 1;
-
-      if (wait >= WAIT_TIME)
-      {
-        motorTestState[A] = ADDING_HALVE;
-        wait = 0;
-      }
-    }
-    else
-    {
-      esc.writeMicroseconds(micros_prog_a = micros_prog_a - MTR_INCREMENT);
-    }
-    break; // End of subtracting
-
-  case ADDING_HALVE: // 3
-#ifdef DEBUG_MOTOR
-    Serial.println("Reached case Adding_halve");
-#endif
-    if (micros_prog_a == MTR_NEUTRAL)
-    {
-      wait += 1;
-      if (wait >= WAIT_TIME)
-      {
-        motorTestState[A] = NEUTRAL;
-        wait = 0;
-      }
-    }
-    else
-      esc.writeMicroseconds(micros_prog_a = micros_prog_a + MTR_INCREMENT);
-    break; // End of ADDING_HALVE
-
-  default:
-    esc.writeMicroseconds(MTR_NEUTRAL);
-    break;
-  }
-  timer_expired = timer_expired + 1;
-}
-
-void prog_b_timer_handler(void)
-{
-  switch (direction)
-  {
-  case Forward:
-
-    if (micros_prog_b == MTR_MAX_ANTICLOCKWISE)
-    {
-      lcd.clear();
-      lcd.home();
-      lcd.print("Green for");
-      lcd.setCursor(0, 1);
-      lcd.print("backwards");
-
-      do
-      {
-        handleButtons(pButtonStates);
-      } while (buttonStates[1] == false); // Wait until green button has been pressed
-
-      direction = Backward; // Change direction
-    }
-
-    else
-      esc.writeMicroseconds(micros_prog_b = micros_prog_b + MTR_INCREMENT);
-
-    break;
-
-    case Backward:
-    if(micros_prog_b == MTR_MIN_CLOCKWISE)
-      {
-        lcd.clear();
-        lcd.home();
-        lcd.print("Green for");
-        lcd.setCursor(0, 1);
-        lcd.print("forwards");
-        
-      do
-      {
-        handleButtons(pButtonStates);
-      } while (buttonStates[1] == false); // Wait until green button has been pressed
-
-      direction = Forward;
-      }
-      else
-        esc.writeMicroseconds(micros_prog_b = micros_prog_b - MTR_INCREMENT);
-    break;
-
-    default:
-    break;
-    
-    }
 }
 
 /*
